# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API del sistema de notas escolares
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: Listar estudiantes (sólo profesor)
      responses:
        "200":
          description: Lista de estudiantes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Registrar un estudiante (sólo profesor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Estudiante creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{studentId}:
    parameters:
      - in: path
        name: studentId
        required: true
        schema:
          type: integer
    get:
      operationId: getStudent
      tags: [students]
      summary: Obtener un estudiante
      responses:
        "200":
          description: Estudiante
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Actualizar un estudiante (sólo profesor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Estudiante actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Eliminar un estudiante (sólo profesor)
      responses:
        "204":
          description: Estudiante eliminado
...
  /students/{studentId}/grades:
    parameters:
      - in: path
        name: studentId
        required: true
        schema:
          type: integer
    get:
      operationId: getStudentGrades
      tags: [grades]
      summary: Obtener todas las calificaciones de un estudiante con su evaluación
      responses:
        "200":
          description: Lista de calificaciones del estudiante
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GradeWithEvaluation"
...
components:
  schemas:
...
    EvaluationType:
      type: string
      description: Tipo de evaluación
      enum: [parcial, recuperativo, rezagado, examen_final, quiz, tarea]
...
    Student:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
          nullable: true
        grade:
          type: string
          nullable: true
          description: Curso o grado del estudiante (ej. "5to A")
        createdAt:
          type: string
          format: date-time
        averageScore:
          type: number
          nullable: true
          description: Promedio general del estudiante
        evaluationsTaken:
          type: integer
          description: Cantidad de evaluaciones con nota registrada
      required: [id, userId, username, fullName, createdAt, evaluationsTaken]
    CreateStudentBody:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 4
        fullName:
          type: string
          minLength: 1
        email:
          type: string
          nullable: true
        grade:
          type: string
          nullable: true
      required: [username, password, fullName]
    UpdateStudentBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          nullable: true
        grade:
          type: string
          nullable: true
        password:
          type: string
          minLength: 4
          description: Si se envía, reemplaza la contraseña
    Evaluation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          $ref: "#/components/schemas/EvaluationType"
        examDate:
          type: string
          format: date-time
        maxScore:
          type: number
          description: Calificación máxima posible (ej. 100)
        weight:
          type: number
          description: Ponderación en porcentaje
        createdAt:
          type: string
          format: date-time
        gradedCount:
          type: integer
          description: Cantidad de estudiantes ya calificados
        totalStudents:
          type: integer
          description: Cantidad total de estudiantes
        averageScore:
          type: number
          nullable: true
          description: Promedio de las calificaciones registradas
      required:
        - id
        - title
        - type
        - examDate
        - maxScore
        - weight
        - createdAt
        - gradedCount
        - totalStudents
...
    Grade:
      type: object
      properties:
        id:
          type: integer
        evaluationId:
          type: integer
        studentId:
          type: integer
        score:
          type: number
          nullable: true
        comment:
          type: string
          nullable: true
        gradedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, evaluationId, studentId]
...
    GradeWithEvaluation:
      type: object
      properties:
        evaluation:
          $ref: "#/components/schemas/Evaluation"
        grade:
          $ref: "#/components/schemas/Grade"
          nullable: true
      required: [evaluation]
...