# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API del sistema de notas escolares
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  userId: number;
  username: string;
  fullName: string;
  email?: string | null;
  /** Curso o grado del estudiante (ej. "5to A") */
  grade?: string | null;
  createdAt: Date;
  /** Promedio general del estudiante */
  averageScore?: number | null;
  /** Cantidad de evaluaciones con nota registrada */
  evaluationsTaken: number;
}