# lib/api-zod/src/generated/types/updateStudentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API del sistema de notas escolares
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStudentBody {
  fullName?: string;
  email?: string | null;
  grade?: string | null;
  /**
   * Si se envía, reemplaza la contraseña
   * @minLength 4
   */
  password?: string;
}